+++ src/services/__init__.py
@@ -3,7 +3,7 @@
 from src.services.blizzard import BlizzardClient, get_blizzard_client
 from src.services.overfast import OverfastClient
 from src.services.spotify import SpotifyClient, get_spotify_client
-from src.services.youtube import YouTubeAudioClient
+from src.services.youtube import YouTubeAudioClient, preload_cookies
 
 __all__ = [
     "BlizzardClient",
@@ -12,4 +12,5 @@ __all__ = [
     "SpotifyClient",
     "get_spotify_client",
     "YouTubeAudioClient",
+    "preload_cookies",
 ]